requirements.txt
numpy>=1.15.1
pandas>=0.23.4
pyinterval>=1.2.0
matplotlib>=2.2.2
seaborn>=0.9.0
pyaml